Move the file to the new module.
