+++ vector/src/main/java/im/vector/app/features/analytics/DecryptionFailureTracker.kt
@@ -17,8 +17,8 @@
 package im.vector.app.features.analytics
 
 import im.vector.app.core.time.Clock
-import im.vector.app.core.utils.compat.removeIfCompat
 import im.vector.app.features.analytics.plan.Error
+import im.vector.lib.core.utils.compat.removeIfCompat
 import im.vector.lib.core.utils.flow.tickerFlow
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
